Do not autorun Acme demo agent test
We do not have a great solution for automatically installing the
requirements of the components, so skip automatically running the
Acme demo agent test for now.
