+++ .github/workflows/run-tests.yml
@@ -33,9 +33,9 @@ jobs:
         if: matrix.os == 'ubuntu-latest' || matrix.os == 'macos-latest'
         run: |
           export MLFLOW_CONDA_HOME=$CONDA
-          pytest --durations=0
+          pytest --durations=0  --ignore=tests/test_acme_demo.py
       - name: Test with pytest on Windows
         if: matrix.os == 'windows-latest'
         run: |
           $ENV:MLFLOW_CONDA_HOME=$ENV:CONDA
-          pytest --durations=0
+          pytest --durations=0  --ignore=tests/test_acme_demo.py
